persist single categories
