+++ pages/index.js
@@ -3,9 +3,24 @@ import Head from "next/head";
 import { motion as m } from "framer-motion";
 import PageContainer from "../components/PageContainer";
 import GameStartForm from "../components/GameStartForm";
+import { useGame } from "../context/GameContext";
+import randomizeCategories from "../utils/randomizeCategories.js";
+import randomizeSlugs from "../utils/randomizeSlugs.js";
 import styles from "./styles.module.css";
+import { useEffect } from "react";
 
-export default function Home() {
+export default function Home({
+  singleCategories,
+  doubleCategories,
+  singleSlugs,
+  doubleSlugs,
+}) {
+  const { setSingleCategories } = useGame();
+
+  useEffect(() => {
+    setSingleCategories(singleCategories);
+  }, []);
+
   return (
     <>
       <Head>
@@ -25,4 +40,14 @@ export default function Home() {
       </PageContainer>
     </>
   );
+}
+
+export async function getServerSideProps() {
+  const [singleCategories, doubleCategories] = await randomizeCategories();
+  const singleSlugs = await randomizeSlugs(singleCategories);
+  const doubleSlugs = await randomizeSlugs(doubleCategories);
+
+  return {
+    props: { singleCategories, doubleCategories, singleSlugs, doubleSlugs },
+  };
 }
